Major functionality Update
Major improvements when it comes to the app working based on what is coming raw from CS2. Added a function to the server to correct observer_slot numbers. Allplayers and observed are nearly almost finished, and with the information learned through those, the others should be decently easy. Clientv2 uses tailwindcss, Giving material ui a shot with Clientv3. Client should be obsolete.