Admin Panel Visual Update
Removed old code so that it only contains Client and Server. Changed locations of some files and how routes work so that now the HUD opens in a separate tab with no other visuals. Also majorly updated the skeleton of the Admin panel. Not yet responsive or functional but a big visual imporvement.