Reorganized File Structure/ Routing
Reorganized files and folders into correct structure and names. Got rid of extra files as well. Reworked Routing to correctly use react-router-dom. This fixed errors saying there was no set path for a page even if there was. Also added a landing page to give the user the choice of going to the admin panel or straight to the hud.