Matchbar / Match API
Added basic functionality to matchbar such as Team names (given from the game) round, and round time. Further progress has been made to the match api but it is not fully working yet. As of now you can press the start button on a created match and it will send the HUD the match information, but still need to figure out how I want to fully implement it into the hud. Also if the hud is refreshed, the match needs to be stopped and started again.